fix favorite func #33
リクエストURIの変数表記変更、お気に入り情報取得関数の記述箇所変更。
