+++ domain/repository/shop_repository.go
@@ -7,7 +7,5 @@ type ShopRepository interface {
 	FindCommentsCount(uint32) (uint32, error)
 	FindFavoritesCount(uint32) (uint32, error)
 	Save(models.Shop) (models.Shop, error)
-	FindFavorites(uint32) ([]models.Favorite, error)
 	SearchShop(string) (models.Shop, error)
-	FindCommentUser(uint32) (models.User, error)
 }
