+++ infrastructure/persistence/favorite.go
@@ -1,6 +1,7 @@
 package persistence
 
 import (
+	"errors"
 	"merpochi_server/domain/models"
 	"merpochi_server/domain/repository"
 
@@ -16,6 +17,31 @@ type favoritePersistence struct {
 // NewFavoritePersistence favoritePersistence構造体の宣言
 func NewFavoritePersistence(db *gorm.DB) repository.FavoriteRepository {
 	return &favoritePersistence{db}
+}
+
+// 指定した店舗のいいね情報を取得
+func (fp *favoritePersistence) FindAll(sid uint32) ([]models.Favorite, error) {
+	var results []models.Favorite
+
+	done := make(chan bool)
+
+	go func(ch chan<- bool) {
+		defer close(ch)
+		query := fp.db.Debug().Table("shops").
+			Select("favorites.*").
+			Joins("inner join favorites on favorites.shop_id = shops.id").
+			Where("shops.id = ?", sid)
+		query.Scan(&results)
+		if len(results) == 0 {
+			ch <- false
+			return
+		}
+		ch <- true
+	}(done)
+	if channels.OK(done) {
+		return results, nil
+	}
+	return []models.Favorite{}, errors.New("no favorite")
 }
 
 // Save お気に入り登録
