+++ infrastructure/router/routes/shops_routes.go
@@ -27,12 +27,6 @@ func iniShopsRoutes() []Route {
 			Handler:      shopHandler.HandleShopCreate,
 			AuthRequired: false,
 		},
-		{
-			URI:          "/shops/{id}/favorites",
-			Method:       http.MethodGet,
-			Handler:      shopHandler.HandleShopFavoritesGet,
-			AuthRequired: false,
-		},
 	}
 	return shopsRoutes
 }
