+++ interfaces/handler/favorite.go
@@ -13,6 +13,7 @@ import (
 
 // FavoriteHandler Userに対するHandlerのインターフェイス
 type FavoriteHandler interface {
+	HandleFavoritesGet(w http.ResponseWriter, r *http.Request)
 	HandleFavoriteCreate(w http.ResponseWriter, r *http.Request)
 	HandleFavoriteDelete(w http.ResponseWriter, r *http.Request)
 }
@@ -28,12 +29,30 @@ func NewFavoriteHandler(fu usecase.FavoriteUsecase) FavoriteHandler {
 	}
 }
 
+// HandleFavoriteGet 指定の店舗に紐付くお気に入り情報を取得
+func (fh favoriteHandler) HandleFavoritesGet(w http.ResponseWriter, r *http.Request) {
+	vars := mux.Vars(r)
+
+	sid, err := strconv.ParseUint(vars["shopId"], 10, 32)
+	if err != nil {
+		responses.ERROR(w, http.StatusBadRequest, err)
+		return
+	}
+
+	favorites, err := fh.favoriteUsecase.GetFavorites(uint32(sid))
+	if err != nil {
+		responses.ERROR(w, http.StatusInternalServerError, err)
+		return
+	}
+	responses.JSON(w, http.StatusOK, favorites)
+}
+
 // HandleFavoriteCreate お気に入りを登録
 func (fh favoriteHandler) HandleFavoriteCreate(w http.ResponseWriter, r *http.Request) {
 	// お気に入りした店舗情報ページのID取得
 	vars := mux.Vars(r)
 
-	sid, err := strconv.ParseUint(vars["id"], 10, 32)
+	sid, err := strconv.ParseUint(vars["shopId"], 10, 32)
 	if err != nil {
 		responses.ERROR(w, http.StatusBadRequest, err)
 		return
@@ -66,7 +85,7 @@ func (fh favoriteHandler) HandleFavoriteCreate(w http.ResponseWriter, r *http.Re
 func (fh favoriteHandler) HandleFavoriteDelete(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 
-	sid, err := strconv.ParseUint(vars["id"], 10, 32)
+	sid, err := strconv.ParseUint(vars["shopId"], 10, 32)
 	if err != nil {
 		responses.ERROR(w, http.StatusBadRequest, err)
 		return
