+++ interfaces/handler/shop.go
@@ -7,16 +7,12 @@ import (
 	"merpochi_server/interfaces/responses"
 	"merpochi_server/usecase"
 	"net/http"
-	"strconv"
-
-	"github.com/gorilla/mux"
 )
 
 // ShopHandler Shopに対するHandlerのインターフェイス
 type ShopHandler interface {
 	HandleShopsGet(w http.ResponseWriter, r *http.Request)
 	HandleShopCreate(w http.ResponseWriter, r *http.Request)
-	HandleShopFavoritesGet(w http.ResponseWriter, r *http.Request)
 }
 
 type shopHandler struct {
@@ -74,22 +70,4 @@ func (sh shopHandler) HandleShopCreate(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	responses.JSON(w, http.StatusCreated, shop)
-}
-
-// HandleShopFavoriteGet 店舗情報を1件取得
-func (sh shopHandler) HandleShopFavoritesGet(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-
-	sid, err := strconv.ParseUint(vars["id"], 10, 32)
-	if err != nil {
-		responses.ERROR(w, http.StatusBadRequest, err)
-		return
-	}
-
-	favorites, err := sh.shopUsecase.GetShopFavorites(uint32(sid))
-	if err != nil {
-		responses.ERROR(w, http.StatusInternalServerError, err)
-		return
-	}
-	responses.JSON(w, http.StatusOK, favorites)
 }
