+++ usecase/favorite.go
@@ -7,6 +7,7 @@ import (
 
 // FavoriteUsecase Favoriteに対するUsecaseのインターフェイス
 type FavoriteUsecase interface {
+	GetFavorites(uint32) ([]models.Favorite, error)
 	CreateFavorite(uint32, uint32) (models.Favorite, error)
 	DeleteFavorite(uint32, uint32) error
 }
@@ -20,6 +21,25 @@ func NewFavoriteUsecase(fr repository.FavoriteRepository) FavoriteUsecase {
 	return &favoriteUsecase{
 		favoriteRepository: fr,
 	}
+}
+
+func (fu favoriteUsecase) GetFavorites(sid uint32) ([]models.Favorite, error) {
+	favorites, err := fu.favoriteRepository.FindAll(sid)
+	if err != nil {
+		return []models.Favorite{}, err
+	}
+	// お気に入りが存在する場合
+	if len(favorites) > 0 {
+		// 取得した店舗のお気に入りに紐付くユーザーを取得
+		for i := 0; i < len(favorites); i++ {
+			favoriteUser, err := fu.favoriteRepository.FindFavoriteUser(favorites[i].UserID)
+			if err != nil {
+				return []models.Favorite{}, err
+			}
+			favorites[i].User = favoriteUser
+		}
+	}
+	return favorites, nil
 }
 
 func (fu favoriteUsecase) CreateFavorite(sid uint32, uid uint32) (models.Favorite, error) {
